+++ jaxrs/src/main/java/de/taimos/dvalin/jaxrs/AbstractAPI.java
@@ -9,9 +9,9 @@ package de.taimos.dvalin.jaxrs;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,109 +20,74 @@ package de.taimos.dvalin.jaxrs;
  * #L%
  */
 
-import java.io.UnsupportedEncodingException;
-import java.net.URI;
-import java.net.URLEncoder;
 import java.util.UUID;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.ws.rs.InternalServerErrorException;
-import javax.ws.rs.RedirectionException;
-import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.SecurityContext;
 
-import de.taimos.dvalin.jaxrs.security.SecurityContextUtil;
-import org.apache.cxf.jaxrs.ext.MessageContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.Autowired;
 
-import com.google.common.base.Charsets;
+import de.taimos.dvalin.jaxrs.context.DvalinRSContext;
 
-public class AbstractAPI implements IContextAware {
-
-    protected MessageContext context;
-    protected HttpServletRequest request;
-    protected HttpServletResponse response;
-
-    @Value("${server.url:http://localhost:${jaxrs.bindport:${svc.port:8080}}}")
-    private String serverURL;
-
+/**
+ * @deprecated inject {@link DvalinRSContext} to replace super methods
+ */
+@Deprecated
+public class AbstractAPI {
+    
+    @Autowired
+    protected DvalinRSContext dvalinContext;
+    
     protected final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-
-    @Override
-    public void setMessageContext(MessageContext context) {
-        this.context = context;
-    }
-
-    @Override
-    public void setHttpServletRequest(HttpServletRequest request) {
-        this.request = request;
-    }
-
-    @Override
-    public void setHttpServletResponse(HttpServletResponse response) {
-        this.response = response;
-    }
-
+    
     @Deprecated
     protected final SecurityContext getSC() {
-        return SecurityContextUtil.getSC();
+        return this.dvalinContext.getSC();
     }
-
+    
     @Deprecated
     protected final void assertSC() {
-        SecurityContextUtil.assertSC();
+        this.dvalinContext.assertLoggedIn();
     }
-
+    
     @Deprecated
     protected final String getUser() {
-        return SecurityContextUtil.getUser();
+        return this.dvalinContext.getCurrentUser().getUsername();
     }
-
+    
     @Deprecated
     protected final boolean hasRole(String role) {
-        return SecurityContextUtil.hasRole(role);
+        return this.dvalinContext.hasRole(role);
     }
-
+    
     @Deprecated
     protected final UUID requestId() {
-        return SecurityContextUtil.requestId();
+        return this.dvalinContext.getRequestId();
     }
-
+    
     protected final String getFirstHeader(String name) {
-        return this.request.getHeader(name);
+        return this.dvalinContext.getFirstHeader(name);
     }
-
+    
     protected final void redirectPath(String path) {
-        this.redirect(this.getServerURL() + path);
+        this.dvalinContext.redirectPath(path);
     }
-
-    protected String getServerURL() {
-        return this.serverURL;
+    
+    protected final String getServerURL() {
+        return this.dvalinContext.getServerURL();
     }
-
+    
     protected final void redirect(String uriString) {
-        throw new RedirectionException(Status.SEE_OTHER, URI.create(uriString));
+        this.dvalinContext.redirect(uriString);
     }
-
-    protected String getCurrentURIEncoded() {
-        try {
-            return URLEncoder.encode(this.getCurrentURI(), Charsets.UTF_8.name());
-        } catch (UnsupportedEncodingException e) {
-            throw new InternalServerErrorException(e);
-        }
+    
+    protected final String getCurrentURIEncoded() {
+        return this.dvalinContext.getCurrentURIEncoded();
     }
-
-    protected String getCurrentURI() {
-        final String path = this.context.getHttpServletRequest().getRequestURI();
-        final String query = this.context.getHttpServletRequest().getQueryString();
-        if (query != null) {
-            return this.getServerURL() + path + "?" + query;
-        }
-        return this.getServerURL() + path;
+    
+    protected final String getCurrentURI() {
+        return this.dvalinContext.getCurrentURI();
     }
-
+    
 }